Separated inference type into multiple files.
